requirements.txt
opencv-python==4.2.0.34
Pillow==7.0.0
torch==1.5.0
torchvision==0.6.0a0+82fd1c8